OCR.bat
The OCR.bat file is a batch script included with the OCR tool that automates the process of running the OCR application on Windows. It streamlines the workflow by allowing you to execute the OCR tool with just a double-click, eliminating the need to manually type commands in the command prompt.